Refactor : 예외 처리 로직들을 헬퍼 메서드로 추출
