+++ src/main/java/com/nbc/newsfeeds/domain/friend/entity/FriendshipStatus.java
@@ -1,4 +1,4 @@
-package com.nbc.newsfeeds.domain.friend.constant;
+package com.nbc.newsfeeds.domain.friend.entity;
 
 public enum FriendshipStatus {
 	PENDING, ACCEPTED, DECLINED, DELETED, CANCELLED
