+++ src/main/java/com/nbc/newsfeeds/domain/friend/service/FriendService.java
@@ -7,8 +7,8 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.nbc.newsfeeds.domain.friend.constant.FriendshipStatus;
 import com.nbc.newsfeeds.domain.friend.entity.Friendship;
+import com.nbc.newsfeeds.domain.friend.entity.FriendshipStatus;
 import com.nbc.newsfeeds.domain.friend.model.request.CursorPageRequest;
 import com.nbc.newsfeeds.domain.friend.model.request.RequestFriendRequest;
 import com.nbc.newsfeeds.domain.friend.model.request.RespondToFriendRequest;
@@ -30,15 +30,8 @@ public class FriendService {
 
 	@Transactional
 	public void requestFriend(Long memberId, RequestFriendRequest req) {
-		if (Objects.equals(memberId, req.targetMemberId())) {
-			// todo 400 BAD_REQUEST
-			throw new RuntimeException("자기 자신과 친구가 될 수 없습니다.");
-		}
-
-		if (friendRepository.existsByMemberIdAndFriendId(memberId, req.targetMemberId())) {
-			// todo 409 CONFLICT
-			throw new RuntimeException("이미 신청이 되어있거나 친구인 상태입니다.");
-		}
+		validateNotSelfRequest(memberId, req.targetMemberId());
+		validateFriendRequestNotExists(memberId, req.targetMemberId());
 
 		Friendship friendship = Friendship.builder()
 			.memberId(memberId)
@@ -50,19 +43,10 @@ public class FriendService {
 
 	@Transactional
 	public void respondToFriendRequest(Long memberId, Long friendshipId, RespondToFriendRequest req) {
-		// todo 404 NOT_FOUND
-		Friendship friendship = friendRepository.findById(friendshipId)
-			.orElseThrow(() -> new RuntimeException("존재하지 않는 친구 요청입니다."));
+		Friendship friendship = getFriendshipOrThrow(friendshipId);
 
-		if (!Objects.equals(friendship.getMemberId(), memberId)) {
-			// todo 403 FORBIDDEN
-			throw new RuntimeException("본인의 친구 요청이 아닙니다.");
-		}
-
-		if (!Objects.equals(friendship.getStatus(), FriendshipStatus.PENDING)) {
-			// todo 409 CONFLICT
-			throw new RuntimeException("이미 처리된 친구 요청입니다.");
-		}
+		validateRequestOwnership(memberId, friendship);
+		validateRequestIsPending(friendship);
 
 		friendship.updateStatus(req.status());
 		friendRepository.save(friendship);
@@ -70,21 +54,10 @@ public class FriendService {
 
 	@Transactional
 	public void deleteFriend(Long memberId, Long friendshipId) {
-		// todo 404 NOT_FOUND
-		Friendship friendship = friendRepository.findById(friendshipId)
-			.orElseThrow(() -> new RuntimeException("존재하지 않는 친구 요청입니다."));
+		Friendship friendship = getFriendshipOrThrow(friendshipId);
 
-		if (!Objects.equals(friendship.getMemberId(), memberId)
-			&& !Objects.equals(friendship.getFriendId(), memberId)
-		) {
-			// todo 403 FORBIDDEN
-			throw new RuntimeException("본인의 친구가 아닙니다.");
-		}
-
-		if (!Objects.equals(friendship.getStatus(), FriendshipStatus.ACCEPTED)) {
-			// todo 409 CONFLICT
-			throw new RuntimeException("본인의 친구가 아닙니다.");
-		}
+		validateRequestOwnership(memberId, friendship);
+		validateRequestIsAccepted(friendship);
 
 		friendship.updateStatus(FriendshipStatus.DELETED);
 		friendRepository.save(friendship);
@@ -130,21 +103,54 @@ public class FriendService {
 
 	@Transactional
 	public void cancelFriendRequest(Long memberId, Long friendshipId) {
-		// todo 404 NOT_FOUND
-		Friendship friendship = friendRepository.findById(friendshipId)
-			.orElseThrow(() -> new RuntimeException("존재하지 않는 친구 요청입니다."));
+		Friendship friendship = getFriendshipOrThrow(friendshipId);
 
-		if (!Objects.equals(friendship.getMemberId(), memberId)) {
-			// todo 403 FORBIDDEN
-			throw new RuntimeException("본인의 친구 요청이 아닙니다.");
-		}
-
-		if (!Objects.equals(friendship.getStatus(), FriendshipStatus.PENDING)) {
-			// todo 409 CONFLICT
-			throw new RuntimeException("이미 처리된 친구 요청입니다.");
-		}
+		validateRequestOwnership(memberId, friendship);
+		validateRequestIsPending(friendship);
 
 		friendship.updateStatus(FriendshipStatus.CANCELLED);
 		friendRepository.save(friendship);
 	}
+
+	private Friendship getFriendshipOrThrow(Long friendshipId) {
+		// todo 404 NOT_FOUND
+		return friendRepository.findById(friendshipId)
+			.orElseThrow(() -> new RuntimeException("존재하지 않는 친구 요청입니다."));
+	}
+
+	private void validateNotSelfRequest(Long memberId, Long targetMemberId) {
+		if (Objects.equals(memberId, targetMemberId)) {
+			// todo 400 BAD_REQUEST
+			throw new RuntimeException("자기 자신과 친구가 될 수 없습니다.");
+		}
+	}
+
+	private void validateFriendRequestNotExists(Long memberId, Long targetMemberId) {
+		if (friendRepository.existsByMemberIdAndFriendId(memberId, targetMemberId)) {
+			// todo 409 CONFLICT
+			throw new RuntimeException("이미 신청이 되어있거나 친구인 상태입니다.");
+		}
+	}
+
+	private void validateRequestOwnership(Long memberId, Friendship friendship) {
+		if (!Objects.equals(friendship.getMemberId(), memberId)) {
+			// todo 403 FORBIDDEN
+			throw new RuntimeException("본인의 친구 정보가 아닙니다.");
+		}
+	}
+
+	private void validateRequestIsPending(Friendship friendship) {
+		if (!Objects.equals(friendship.getStatus(), FriendshipStatus.PENDING)) {
+			// todo 409 CONFLICT
+			throw new RuntimeException("이미 처리된 친구 정보가 입니다.");
+		}
+	}
+
+	private void validateRequestIsAccepted(Friendship friendship) {
+		if (!Objects.equals(friendship.getStatus(), FriendshipStatus.ACCEPTED)) {
+			// todo 409 CONFLICT
+			throw new RuntimeException("본인의 친구가 아닙니다.");
+		}
+	}
+
 }
